rust_runtime: Move cpp wrapper to sdk dir
